@@ -1,6 +1,6 @@
-package Repository;
+package repository;
 
-import Entity.FunctionTypeEntity;
+import entity.FunctionTypeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
