@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
-@EntityScan("Entity")
-@EnableJpaRepositories("Repository")
+@EntityScan("entity")
+@EnableJpaRepositories("repository")
 public class Application {
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
