@@ -1,7 +1,7 @@
 package service;
 
-import Repository.FunctionTypeRepository;
-import Entity.FunctionTypeEntity;
+import repository.FunctionTypeRepository;
+import entity.FunctionTypeEntity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
