@@ -1,7 +1,10 @@
 package service;
 
-import Repository.UserRepository;
-import Entity.UserEntity;
+import dto.UserResponse;
+import org.mapstruct.Mapper;
+import org.mapstruct.factory.Mappers;
+import repository.UserRepository;
+import entity.UserEntity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +14,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import org.springframework.security.crypto.password.PasswordEncoder;
+import java.time.Instant;
+
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -47,7 +54,6 @@ public class UserService {
 
     // Множественный поиск
     public List<UserEntity> findUsersByUsernameFragment(String usernameFragment) {
-        logger.info("Поиск пользователей по фрагменту имени: {}", usernameFragment);
         List<UserEntity> users = userRepository.findByUsernameContainingIgnoreCase(usernameFragment);
         logger.debug("Найдено {} пользователей по фрагменту имени: '{}'", users.size(), usernameFragment);
         return users;
@@ -63,7 +69,6 @@ public class UserService {
 
     // поиск с пагинацией
     public Page<UserEntity> findAllUsersPaged(int page, int size, String sortBy, String direction) {
-        logger.info("Получение всех пользователей с пагинацией (страница: {}, размер: {}) и сортировкой по '{} {}'", page, size, sortBy, direction);
         Sort sort = direction.equalsIgnoreCase("desc") ? Sort.by(sortBy).descending() : Sort.by(sortBy).ascending();
         Pageable pageable = PageRequest.of(page, size, sort);
         Page<UserEntity> usersPage = userRepository.findAll(pageable);
@@ -78,5 +83,112 @@ public class UserService {
         List<UserEntity> users = userRepository.findUsersByFunctionTypeId(typeId);
         logger.debug("Найдено {} пользователей, у которых есть функции типа ID: {}", users.size(), typeId);
         return users;
+    }
+
+    private final PasswordEncoder passwordEncoder;
+
+    @Autowired
+    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
+        this.userRepository = userRepository;
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    // <<<>>> маппинг
+    @Mapper
+    public interface UserMapper {
+        UserMapper INSTANCE = Mappers.getMapper(UserMapper.class);
+
+        UserResponse toUserResponse(UserEntity entity);
+    }
+
+    // 1. Регистрация
+    public UserEntity register(String email, String username, String password) {
+        if (userRepository.findByUsername(username).isPresent()) {
+            throw new IllegalArgumentException("Пользователь с именем '" + username + "' уже существует");
+        }
+        if (userRepository.findByEmail(email).isPresent()) {
+            throw new IllegalArgumentException("Пользователь с email '" + email + "' уже существует");
+        }
+
+        // Хеширование пароля
+        String encodedPassword = passwordEncoder.encode(password);
+
+        UserEntity user = new UserEntity();
+        user.setEmail(email);
+        user.setUsername(username);
+        user.setPasswordHash(encodedPassword);
+        user.setRole("USER");
+        user.setCreatedAt(LocalDateTime.from(Instant.now()));
+
+        // Сохранение
+        UserEntity saved = userRepository.save(user);
+        logger.info("Пользователь ID={} успешно зарегистрирован", saved.getId());
+        return saved;
+    }
+
+    // 2. Аутентификация (логин)
+    public UserEntity authenticate(String usernameOrEmail, String rawPassword) {
+        Optional<UserEntity> userOpt = userRepository.findByUsernameOrEmail(usernameOrEmail);
+
+        if (userOpt.isEmpty()) {
+            throw new IllegalArgumentException("Неверные учётные данные");
+        }
+
+        UserEntity user = userOpt.get();
+
+        // Сравнение пароля с хешем
+        if (!passwordEncoder.matches(rawPassword, user.getPasswordHash())) {
+            throw new IllegalArgumentException("Неверные учётные данные");
+        }
+
+        logger.info("Пользователь ID={} успешно вошёл", user.getId());
+        return user;
+    }
+    // <<<<>>>>
+    // Для тестирования: эмулируем, что пользователь уже залогинен и идентифицирован
+    // В реальной версии замените на SecurityContextHolder или JWT-парсинг
+    private static final ThreadLocal<Long> CURRENT_USER_ID = new ThreadLocal<>();
+
+    // Метод для тестового входа (используется в @PostConstruct или тестах)
+    public void setCurrentUserIdForTesting(Long userId) {
+        CURRENT_USER_ID.set(userId);
+    }
+
+    // 3. Получение текущего пользователя (временно)
+    public UserResponse getCurrentUser() {
+        Long userId = CURRENT_USER_ID.get();
+        if (userId == null) {
+            // Вариант 1: выбрасывать — строгий режим
+            // throw new IllegalArgumentException("Пользователь не авторизован");
+
+            // Вариант 2: возвращать демо-пользователя (для UI-тестов)
+            // Создаём временного пользователя, если его ещё нет
+            String demoEmail = "demo@example.com";
+            Optional<UserEntity> demoOpt = userRepository.findByEmail(demoEmail);
+            UserEntity demo = demoOpt.orElseGet(() -> {
+                UserEntity u = new UserEntity();
+                u.setEmail(demoEmail);
+                u.setUsername("demo_user");
+                u.setPasswordHash(passwordEncoder.encode("demo123"));
+                u.setRole("USER");
+                u.setCreatedAt(LocalDateTime.now());
+                return userRepository.save(u);
+            });
+            return toUserResponse(demo);
+        }
+
+        return toUserResponse(
+                userRepository.findById(userId)
+                        .orElseThrow(() -> new IllegalArgumentException("Пользователь не найден"))
+        );
+    }
+    private UserResponse toUserResponse(UserEntity e) {
+        UserResponse dto = new UserResponse();
+        dto.setId(e.getId());
+        dto.setUsername(e.getUsername());
+        dto.setEmail(e.getEmail());
+        dto.setRole(e.getRole());
+        dto.setCreatedAt(e.getCreatedAt());
+        return dto;
     }
 }
